tidy: refactor to reduce complexity
