+++ src/Cli/RunCommand.php
@@ -2,7 +2,6 @@
 namespace GT\Cron\Cli;
 
 use DateTime;
-use DateTimeZone;
 use Gt\Cli\Argument\ArgumentValueList;
 use Gt\Cli\Command\Command;
 use Gt\Cli\Parameter\NamedParameter;
@@ -20,9 +19,11 @@ use GT\Cron\RunnerFactory;
 use GT\Cron\ScriptExecutionException;
 
 class RunCommand extends Command {
+	private ?LocalTime $localTime = null;
+
 	/** @SuppressWarnings(PHPMD.ExitExpression) */
 	public function run(?ArgumentValueList $arguments = null):int {
-		$this->applySystemTimezone();
+		$this->localTime()->applySystemTimezone();
 
 		$filename = $arguments->get("file", "crontab");
 		$filePath = implode(DIRECTORY_SEPARATOR, [
@@ -152,7 +153,7 @@ class RunCommand extends Command {
 		?string $nextCommand = null
 	):void {
 		$now = new DateTime();
-		$this->stream->writeLine("Current time: " . $this->formatLocalTime($now));
+		$this->stream->writeLine("Current time: " . $this->localTime()->format($now));
 
 		if(is_null($wait)) {
 			$this->writeLine("No tasks in crontab.");
@@ -174,7 +175,7 @@ class RunCommand extends Command {
 
 		$this->stream->writeLine($message);
 
-		$message = "Next job at: " . $this->formatLocalTime($wait);
+		$message = "Next job at: " . $this->localTime()->format($wait);
 		if($nextCommand) {
 			$message .= " [" . $this->displayCommandName($nextCommand) . "]";
 		}
@@ -232,79 +233,12 @@ class RunCommand extends Command {
 		}
 	}
 
-	protected function applySystemTimezone():void {
-		if($timezone = $this->detectSystemTimezone()) {
-			date_default_timezone_set($timezone);
-		}
-	}
-
-	protected function detectSystemTimezone():?string {
-		return $this->detectTimezoneFromEnvironment()
-			?? $this->detectTimezoneFromLocaltime()
-			?? $this->detectTimezoneFromTimezoneFile();
-	}
-
-	protected function detectTimezoneFromEnvironment():?string {
-		$environmentTimezone = getenv("TZ");
-		if($environmentTimezone !== false
-		&& $this->isValidTimezone($environmentTimezone)) {
-			return $environmentTimezone;
+	private function localTime():LocalTime {
+		if(is_null($this->localTime)) {
+			$this->localTime = new LocalTime();
 		}
 
-		return null;
-	}
-
-	protected function detectTimezoneFromLocaltime():?string {
-		$localtimePath = "/etc/localtime";
-		if(is_link($localtimePath)) {
-			$link = readlink($localtimePath);
-			if($link !== false
-			&& preg_match("#/zoneinfo/(.+)$#", $link, $match)
-			&& $this->isValidTimezone($match[1])) {
-				return $match[1];
-			}
-		}
-
-		return null;
-	}
-
-	protected function detectTimezoneFromTimezoneFile():?string {
-		$timezonePath = "/etc/timezone";
-		if(is_file($timezonePath)) {
-			$timezone = file_get_contents($timezonePath);
-			if($timezone === false) {
-				return null;
-			}
-
-			$timezone = trim($timezone);
-			if($this->isValidTimezone($timezone)) {
-				return $timezone;
-			}
-		}
-
-		return null;
-	}
-
-	protected function isValidTimezone(string $timezone):bool {
-		return in_array(
-			$timezone,
-			DateTimeZone::listIdentifiers(),
-			true
-		);
-	}
-
-	protected function formatLocalTime(DateTime $dateTime):string {
-		$local = clone $dateTime;
-		$local->setTimezone(new DateTimeZone(date_default_timezone_get()));
-		$message = $local->format("H:i:s");
-
-		if($local->getOffset() !== 0) {
-			$utc = clone $local;
-			$utc->setTimezone(new DateTimeZone("UTC"));
-			$message .= " (" . $utc->format("H:i:s") . " UTC)";
-		}
-
-		return $message;
+		return $this->localTime;
 	}
 
 	public function getName():string {
